Added Menu screen and buttons
I added a menu screen with buttons, allowing the menu to be pulled up again by pressing the letter "M" even after starting the game. I also removed the camera and decided to implement levels or zones instead of a moving background that follows the player, due to time constraints and bugs that were occurring.